third_party/WebKit/LayoutTests/http/tests/inspector/console/console-links-in-errors-with-trace.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../console-test.js"></script>
<script src="../resources/source3.js"></script>
<script>

function test()
{
    InspectorTest.evaluateInPage("foo()", step1);
    WebInspector.inspectorView._tabbedPane.addEventListener(WebInspector.TabbedPane.Events.TabSelected, panelChanged);

    function panelChanged()
    {
        InspectorTest.addResult("Panel " + WebInspector.inspectorView._tabbedPane._currentTab.id + " was opened.");
        InspectorTest.completeTest();
    }

    var clickTarget;
    function step1()
    {
        var firstMessageEl = WebInspector.ConsoleView.instance()._visibleViewMessages[0].element();
        clickTarget = firstMessageEl.querySelectorAll(".console-message-text a")[1];
        WebInspector.inspectorView.showPanel("console").then(testClickTarget);
    }

    function testClickTarget()
    {
        InspectorTest.addResult("Clicking link " + clickTarget.textContent);
        clickTarget.click();
    }

    InspectorFrontendHost.openInNewTab = function()
    {
        InspectorTest.addResult("Failure: Open link in new tab!!");
        InspectorTest.completeTest();
    };
}
</script>
</head>
<body onload="runTest()">
<p>
Test that relative links in traces open in the sources panel.
</p>
</body>
</html>
