third_party/WebKit/LayoutTests/http/tests/inspector/console/console-links-in-errors-with-trace-expected.txt

CONSOLE ERROR: line 3: Error: http://127.0.0.1:8000/inspector/resources/source3.js
Test that relative links in traces open in the sources panel.

Clicking link source3.js:3
Panel sources was opened.
